chore: accordion placement in marketing and components
placed old accordion in marketing and accordian in components as accordion closes #313
