+++ src/lib/marketing/index.ts
@@ -1,3 +1,5 @@
+export { default as Accordion } from './DetailsAccordion/DetailsAccordion.svelte';
+
 export { default as Header } from './Header/Header.svelte';
 
 export { default as Container } from './Container/Container.svelte';
